+++ iconservice/iiss/engine.py
@@ -14,18 +14,19 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from typing import TYPE_CHECKING, Any
+from typing import TYPE_CHECKING, Any, Optional, List
 
 from iconcommons.logger import Logger
-from .reward_calc.ipc.reward_calc_proxy import RewardCalcProxy
-from iconservice.iiss.reward_calc.ipc.message import CalculateResponse
-from .commit_delegator import CommitDelegator
-from .handler.delegation_handler import DelegationHandler
-from .handler.iscore_handler import IScoreHandler
-from .handler.stake_handler import StakeHandler
 from .reward_calc.data_creator import DataCreator as RewardCalcDataCreator
-from .reward_calc.msg_data import PRepUnregisterTx
-from ..icon_constant import IISS_SOCKET_PATH
+from .reward_calc.ipc.message import CalculateResponse
+from .reward_calc.ipc.reward_calc_proxy import RewardCalcProxy
+from ..base.ComponentBase import EngineBase
+from ..base.exception import InvalidParamsException
+from ..base.type_converter import TypeConverter
+from ..base.type_converter_templates import ConstantKeys, ParamType
+from ..icon_constant import IISS_SOCKET_PATH, IISS_MAX_DELEGATIONS, I_SCORE_EXCHANGE_RATE
+from ..iconscore.icon_score_event_log import EventLogEmitter
+from ..icx import Intent
 from ..iiss.issue_formula import IssueFormula
 
 if TYPE_CHECKING:
@@ -33,23 +34,29 @@ if TYPE_CHECKING:
     from ..iconscore.icon_score_context import IconScoreContext
     from ..precommit_data_manager import PrecommitData
 
-    from ..base.address import Address
-    from .reward_calc.msg_data import PRepRegisterTx, TxData
+    from ..base.address import Address, ZERO_SCORE_ADDRESS
+    from ..icx.icx_account import Account
+    from .reward_calc.msg_data import PRepRegisterTx, PRepUnregisterTx, TxData, DelegationInfo, DelegationTx, Header, \
+        BlockProduceInfoData, PRepsData
+    from .storage import Reward
+    from ..prep.storage import GovernanceVariable
 
 
-class Engine:
+class Engine(EngineBase):
 
     def __init__(self):
+        super().__init__()
+
         self._invoke_handlers: dict = {
-            'setStake': StakeHandler.handle_set_stake,
-            'setDelegation': DelegationHandler.handle_set_delegation,
-            'claimIScore': IScoreHandler.handle_claim_iscore
+            'setStake': self.handle_set_stake,
+            'setDelegation': self.handle_set_delegation,
+            'claimIScore': self.handle_claim_iscore
         }
 
         self._query_handler: dict = {
-            'getStake': StakeHandler.handle_get_stake,
-            'getDelegation': DelegationHandler.handle_get_delegation,
-            'queryIScore': IScoreHandler.handle_query_iscore
+            'getStake': self.handle_get_stake,
+            'getDelegation': self.handle_get_delegation,
+            'queryIScore': self.handle_query_iscore
         }
 
         self._reward_calc_proxy: 'RewardCalcProxy' = None
@@ -57,13 +64,8 @@ class Engine:
 
     def open(self, context: 'IconScoreContext', path: str):
         self._init_reward_calc_proxy(path)
-
-        self._init_commit_delegator()
         # todo: consider formula managing r min, r max, r point
         self._formula = IssueFormula()
-
-        handlers: list = [StakeHandler, DelegationHandler, IScoreHandler]
-        self._init_handlers(handlers)
 
     # TODO implement calculate callback function
     def calculate_callback(self, cb_data: 'CalculateResponse'):
@@ -81,9 +83,6 @@ class Engine:
     def _init_handlers(self, handlers: list):
         for handler in handlers:
             handler.reward_calc_proxy = self._reward_calc_proxy
-
-    def _init_commit_delegator(self):
-        CommitDelegator.reward_calc_proxy = self._reward_calc_proxy
 
     def close(self):
         self._close_reward_calc_proxy()
@@ -104,14 +103,14 @@ class Engine:
         return ret
 
     def genesis_commit(self, context: 'IconScoreContext', precommit_data: 'PrecommitData'):
-        CommitDelegator.genesis_update_db(context, precommit_data)
+        self.genesis_update_db(context, precommit_data)
         context.storage.iiss.commit(precommit_data.rc_block_batch)
-        CommitDelegator.genesis_send_ipc(context, precommit_data)
+        self.genesis_send_ipc(context, precommit_data)
 
     def commit(self, context: 'IconScoreContext', precommit_data: 'PrecommitData'):
-        CommitDelegator.update_db(context, precommit_data)
+        self.update_db(context, precommit_data)
         context.storage.iiss.commit(precommit_data.rc_block_batch)
-        CommitDelegator.send_ipc(context, precommit_data)
+        self.send_ipc(context, precommit_data)
 
     def create_icx_issue_info(self, context: 'IconScoreContext'):
         gv: 'GovernanceVariable' = context.storage.prep.get_gv(context)
@@ -132,26 +131,315 @@ class Engine:
     def rollback(self):
         pass
 
-    # TODO we don't allow inner function except these functions
-    def put_reg_prep_candidate_for_rc_data(self,
-                                           batch: list,
-                                           address: 'Address',
-                                           block_height: int):
-        tx: 'PRepRegisterTx' = RewardCalcDataCreator.create_tx_prep_reg()
-        iiss_tx_data: 'TxData' = RewardCalcDataCreator.create_tx(address, block_height, tx)
-        self._rc_storage.storage.rc.put(batch, iiss_tx_data)
+    # ========================
+    def handle_set_stake(self, context: 'IconScoreContext', params: dict, tx_result: 'TransactionResult'):
 
-    def put_unreg_prep_candidate_for_iiss_db(self,
-                                             batch: list,
-                                             address: 'Address',
-                                             block_height: int):
-        tx: 'PRepUnregisterTx' = RewardCalcDataCreator.create_tx_prep_unreg()
-        iiss_tx_data: 'TxData' = RewardCalcDataCreator.create_tx(address, block_height, tx)
-        self._rc_storage.put(batch, iiss_tx_data)
+        address: 'Address' = context.tx.origin
+        ret_params: dict = TypeConverter.convert(params, ParamType.IISS_SET_STAKE)
+        value: int = ret_params[ConstantKeys.VALUE]
 
-    def apply_candidate_delegated_offset_for_iiss_variable(self,
-                                                           context: 'IconScoreContext',
-                                                           offset: int):
-        total_delegated_amount: int = self._variable.issue.get_total_candidate_delegated(context)
-        self._variable.issue.put_total_candidate_delegated(context,
-                                                           total_delegated_amount + offset)
+        self._put_stake_for_state_db(context, address, value)
+
+    @classmethod
+    def _put_stake_for_state_db(cls, context: 'IconScoreContext', address: 'Address', value: int):
+
+        if not isinstance(value, int) or value < 0:
+            raise InvalidParamsException('Failed to stake: value is not int type or value < 0')
+
+        account: 'Account' = context.storage.icx.get_account(context, address, Intent.STAKE)
+        unstake_lock_period = context.storage.iiss.get_unstake_lock_period(context)
+        account.set_stake(value, unstake_lock_period)
+        context.storage.icx.put_account(context, account)
+        # TODO tx_result make if needs
+
+    def handle_get_stake(self, context: 'IconScoreContext', params: dict) -> dict:
+
+        ret_params: dict = TypeConverter.convert(params, ParamType.IISS_GET_STAKE)
+        address: 'Address' = ret_params[ConstantKeys.ADDRESS]
+        return self._get_stake(context, address)
+
+    @classmethod
+    def _get_stake(cls, context: 'IconScoreContext', address: 'Address') -> dict:
+
+        account: 'Account' = context.storage.icx.get_account(context, address, Intent.STAKE)
+
+        stake: int = account.stake
+        unstake: int = account.unstake
+        unstake_block_beight: int = account.unstake_block_height
+
+        data = {
+            "stake": stake,
+            "unstake": unstake,
+            "unstakedBlockHeight": unstake_block_beight
+        }
+        return data
+
+    def handle_set_delegation(self, context: 'IconScoreContext', params: dict, tx_result: 'TransactionResult'):
+
+        address: 'Address' = context.tx.origin
+        ret_params: dict = TypeConverter.convert(params, ParamType.IISS_SET_DELEGATION)
+        data: list = ret_params[ConstantKeys.DELEGATIONS]
+
+        self._put_delegation_for_state_db(context, address, data)
+        self._put_delegation_for_iiss_db(context, address, data)
+
+    @classmethod
+    def _put_delegation_for_state_db(cls,
+                                     context: 'IconScoreContext',
+                                     delegating_address: 'Address',
+                                     delegations: list):
+
+        if not isinstance(delegations, list):
+            raise InvalidParamsException('Failed to delegation: delegations is not list type')
+
+        if len(delegations) > IISS_MAX_DELEGATIONS:
+            raise InvalidParamsException(f'Failed to delegation: Overflow Max Input List')
+
+        delegating: 'Account' = context.storage.icx.get_account(context, delegating_address, Intent.DELEGATING)
+        candidates: dict = cls._make_candidates(delegating.delegations, delegations)
+        cls._set_delegations(delegating, candidates)
+        dirty_accounts: list = cls._delegated_candidates(context, delegating, candidates)
+        dirty_accounts.append(delegating)
+
+        for dirty_account in dirty_accounts:
+            context.storage.icx.put_account(context, dirty_account)
+            if context.engine.prep.is_candidate(context, dirty_account.address):
+                context.engine.prep.update_sorted_candidates(context,
+                                                             dirty_account.address,
+                                                             dirty_account.delegated_amount)
+        # TODO tx_result make if needs
+
+    @classmethod
+    def _make_candidates(cls, old_delegations: list, new_delegations: list) -> dict:
+        candidates: dict = {}
+
+        if old_delegations:
+            for address, old in old_delegations:
+                candidates[address] = (old, 0)
+
+        for delegation in new_delegations:
+            address, new = delegation.values()
+            if address in candidates:
+                old, _ = candidates[address]
+                candidates[address] = (old, new)
+            else:
+                candidates[address] = (0, new)
+        return candidates
+
+    @classmethod
+    def _set_delegations(cls, delegating: 'Account', candidates: dict):
+        new_delegations: list = [(address, new) for address, (before, new) in candidates.items() if new > 0]
+        delegating.set_delegations(new_delegations)
+
+        if delegating.delegations_amount > delegating.stake:
+            raise InvalidParamsException(
+                f"Failed to delegation: delegation_amount{delegating.delegations_amount} > stake{delegating.stake}")
+
+    @classmethod
+    def _delegated_candidates(cls, context: 'IconScoreContext', delegating: 'Account', candidates: dict) -> list:
+        dirty_accounts: list = []
+        for address, (before, new) in candidates.items():
+            if address == delegating.address:
+                candidate: 'Account' = delegating
+            else:
+                candidate: 'Account' = context.storage.icx.get_account(context, address, Intent.DELEGATED)
+                dirty_accounts.append(candidate)
+
+            offset: int = new - before
+            candidate.update_delegated_amount(offset)
+
+            if context.engine.prep.is_candidate(context, address):
+                total: int = context.storage.prep.get_total_candidate_delegated(context)
+                context.storage.prep.put_total_candidate_delegated(context, total + offset)
+        return dirty_accounts
+
+    @classmethod
+    def _put_delegation_for_iiss_db(cls, context: 'IconScoreContext', address: 'Address', delegations: list):
+
+        delegation_list: list = []
+
+        for delegation in delegations:
+            delegation_address, delegation_value = delegation.values()
+            info: 'DelegationInfo' = RewardCalcDataCreator.create_delegation_info(delegation_address, delegation_value)
+            delegation_list.append(info)
+
+        delegation_tx: 'DelegationTx' = RewardCalcDataCreator.create_tx_delegation(delegation_list)
+        iiss_tx_data: 'TxData' = RewardCalcDataCreator.create_tx(address, context.block.height, delegation_tx)
+        context.storage.rc.put(context.rc_block_batch, iiss_tx_data)
+
+    def handle_get_delegation(self, context: 'IconScoreContext', params: dict) -> dict:
+
+        ret_params: dict = TypeConverter.convert(params, ParamType.IISS_GET_STAKE)
+        address: 'Address' = ret_params[ConstantKeys.ADDRESS]
+        return self._get_delegation(context, address)
+
+    @classmethod
+    def _get_delegation(cls, context: 'IconScoreContext', address: 'Address') -> dict:
+
+        account: 'Account' = context.storage.icx.get_account(context, address, Intent.DELEGATED)
+        delegation_list: list = []
+        for address, value in account.delegations:
+            delegation_list.append({"address": address, "value": value})
+
+        data = {
+            "delegations": delegation_list,
+            "totalDelegated": account.delegations_amount
+        }
+
+        return data
+
+    @classmethod
+    def _iscore_to_icx(cls, iscore: int) -> int:
+        return iscore // I_SCORE_EXCHANGE_RATE
+
+    def handle_claim_iscore(self, context: 'IconScoreContext', params: dict, tx_result: 'TransactionResult'):
+        address: 'Address' = context.tx.origin
+
+        # TODO: error handling
+        iscore, block_height = self._reward_calc_proxy.claim_iscore(address, context.block.height, context.block.hash)
+
+        icx: int = self._iscore_to_icx(iscore)
+
+        from_account: 'Account' = context.storage.icx.get_account(context, address)
+        from_account.deposit(icx)
+        context.storage.icx.put_account(context, from_account)
+        self._create_tx_result(context, iscore, icx)
+
+    @classmethod
+    def _create_tx_result(cls, context: 'IconScoreContext', iscore: int, icx: int):
+        # make tx result
+        event_signature: str = 'IScoreClaimed(int,int)'
+        arguments = [iscore, icx]
+        index = 0
+        EventLogEmitter.emit_event_log(context, ZERO_SCORE_ADDRESS, event_signature, arguments, index)
+
+    def handle_query_iscore(self, context: 'IconScoreContext', params: dict) -> dict:
+        ret_params: dict = TypeConverter.convert(params, ParamType.IISS_QUERY_ISCORE)
+        address: 'Address' = ret_params[ConstantKeys.ADDRESS]
+
+        # TODO: error handling
+        iscore, block_height = self._reward_calc_proxy.query_iscore(address)
+
+        data = {
+            "iscore": iscore,
+            "icx": self._iscore_to_icx(iscore),
+            "blockHeight": block_height
+        }
+
+        return data
+
+    def genesis_update_db(self, context: 'IconScoreContext', precommit_data: 'PrecommitData'):
+        context.engine.prep.update_preps_to_variable(context)
+        self._put_next_calc_block_height(context, precommit_data.block.height)
+
+        self._put_header_for_rc(context, precommit_data)
+        self._put_gv_for_rc(context, precommit_data)
+        self._put_preps_for_rc(context, precommit_data)
+
+    def genesis_send_ipc(self, context: 'IconScoreContext', precommit_data: 'PrecommitData'):
+        block_height: int = precommit_data.block.height
+        path: str = context.storage.rc.create_db_for_calc(precommit_data.block.height)
+        self._reward_calc_proxy.commit_block(True, block_height, precommit_data.block.hash)
+        self._reward_calc_proxy.calculate(path, block_height)
+
+    def update_db(self, context: 'IconScoreContext', precommit_data: 'PrecommitData'):
+
+        # every block time
+        self._put_block_produce_info_for_rc(context, precommit_data)
+        # cls._put_preps_for_rc(context, precommit_data)
+
+        if not self._check_update_calc_period(context, precommit_data):
+            return
+
+        self._put_header_for_rc(context, precommit_data)
+        self._put_gv_for_rc(context, precommit_data)
+
+    def send_ipc(self, context: 'IconScoreContext', precommit_data: 'PrecommitData'):
+        # TODO: Disable reward_calc_proxy for test
+        return
+
+        # every block
+        self._reward_calc_proxy.commit_block(True, precommit_data.block.height, precommit_data.block.hash)
+
+        if not cls._check_update_calc_period(context, precommit_data):
+            return
+
+        block_height: int = precommit_data.block.height
+        path: str = cls.rc_storage.create_db_for_calc(precommit_data.block.height)
+        cls.reward_calc_proxy.calculate(path, block_height)
+        cls._put_next_calc_block_height(context, precommit_data.block.height)
+
+    @classmethod
+    def _check_update_calc_period(cls, context: 'IconScoreContext', precommit_data: 'PrecommitData') -> bool:
+        block_height: int = precommit_data.block.height
+        check_next_block_height: Optional[int] = context.storage.rc.get_calc_next_block_height(context)
+        if check_next_block_height is None:
+            return False
+
+        return block_height == check_next_block_height
+
+    @classmethod
+    def _put_next_calc_block_height(cls, context: 'IconScoreContext', block_height: int):
+        calc_period: int = context.storage.rc.get_calc_period(context)
+        if calc_period is None:
+            raise InvalidParamsException("Fail put next calc block height: didn't init yet")
+        context.storage.rc.put_calc_next_block_height(context, block_height + calc_period)
+
+    @classmethod
+    def _put_header_for_rc(cls, context: 'IconScoreContext', precommit_data: 'PrecommitData'):
+        data: 'Header' = RewardCalcDataCreator.create_header(0, precommit_data.block.height)
+        context.storage.rc.put(precommit_data.rc_block_batch, data)
+
+    @classmethod
+    def _put_gv_for_rc(cls, context: 'IconScoreContext', precommit_data: 'PrecommitData'):
+        gv: 'GovernanceVariable' = context.storage.get_gv(context)
+
+        current_total_supply = context.storage.icx.get_total_supply(context)
+        current_total_candidate_delegated = context.storage.iiss.get_total_candidate_delegated(context)
+        reward_prep: 'Reward' = context.storage.iiss.get_reward_prep(context)
+
+        reward_rep: int = IssueFormula.calculate_r_rep(reward_prep.reward_min,
+                                                       reward_prep.reward_max,
+                                                       reward_prep.reward_point,
+                                                       current_total_supply,
+                                                       current_total_candidate_delegated)
+        calculated_incentive_rep: int = IssueFormula.calculate_i_rep_per_block_contributor(gv.incentive_rep)
+        reward_prep.reward_rate = reward_rep
+        context.storage.iiss.put_reward_prep(context, reward_prep)
+
+        data: 'GovernanceVariable' = RewardCalcDataCreator.create_gv_variable(precommit_data.block.height,
+                                                                              calculated_incentive_rep,
+                                                                              reward_rep)
+        context.storage.rc.put(precommit_data.rc_block_batch, data)
+
+    @classmethod
+    def _put_block_produce_info_for_rc(cls, context: 'IconScoreContext', precommit_data: 'PrecommitData'):
+        if precommit_data.prev_block_generator is None or precommit_data.prev_block_validators is None:
+            return
+
+        generator: 'Address' = precommit_data.prev_block_generator
+        validators: List['Address'] = precommit_data.prev_block_validators
+
+        Logger.debug(f"put_block_produce_info_for_rc", "iiss")
+        data: 'BlockProduceInfoData' = RewardCalcDataCreator.create_block_produce_info_data(precommit_data.block.height,
+                                                                                            generator,
+                                                                                            validators)
+        context.storage.rc.put(precommit_data.rc_block_batch, data)
+
+    @classmethod
+    def _put_preps_for_rc(cls, context: 'IconScoreContext', precommit_data: 'PrecommitData'):
+        preps: List['PRep'] = context.updated_preps
+
+        if len(preps) == 0:
+            return
+
+        total_candidate_delegated: int = 0
+        for prep in preps:
+            total_candidate_delegated += prep.delegated
+
+        Logger.debug(f"put_preps_for_rc: total_candidate_delegated{total_candidate_delegated}", "iiss")
+
+        data: 'PRepsData' = RewardCalcDataCreator.create_prep_data(precommit_data.block.height,
+                                                                   total_candidate_delegated,
+                                                                   preps)
+        context.storage.rc.put(precommit_data.rc_block_batch, data)
